allergy만 반환(dto 수정)

+++ MBTI-backend/src/main/java/tukorea/MBTIbackend/domain/product/controller/ProductController.java
@@ -14,7 +14,7 @@ public class ProductController {
 
 
     @GetMapping("/api/product/{productId}")
-    public ProductResponseDto findProductById(@PathVariable Integer productId) {
+    public ProductResponseDto findProductById(@PathVariable Long productId) {
         return productService.findProductByPID(productId);
     }
 }
